week-1-2/requirements.txt
braintrust
cohere
kagglehub
lancedb
openai
pandas
pydantic
python-dotenv
rich
typing